pyproject.toml
[tool.poetry]
name = "goit-python-web-hw-14"
version = "0.1.0"
description = ""
authors = ["lexxai <lexxai@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.4.2"}
alembic = "^1.12.1"
jinja2 = "^3.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.36.0"
colorlog = "^6.7.0"
aiohttp = "^3.9.0"
faker = "^20.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.format]
max-line-length = "119"

[tool.black]
line-length = 119